requirements.txt
streamlit
google-generativeai
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-google-genai
PyPDF2
faiss-cpu